pyproject.toml
# NOTE: you have to use single-quoted strings in TOML for regular expressions.
# It's the equivalent of r-strings in Python.  Multiline strings are treated as
# verbose regular expressions by Black.  Use [ ] to denote a significant space
# character.

[tool.poetry]
name = "papermill-origami"
version = "0.0.5"
description = "The noteable API interface"
authors = ["Matt Seal <matt@noteable.io>"]
maintainers = ["Matt Seal <matt@noteable.io>"]
license = "BSD-3-Clause"
readme = "README.md"
repository = "https://github.com/noteable-io/papermill-origami"
# old setup.cfg had a bdist_wheel option.
# To build a wheel, use poetry build -f wheel
keywords = ["notebook", "api", "noteable"]
classifiers=[
    'Intended Audience :: Developers',
    'License :: OSI Approved :: BSD License',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]

# Manifest.in is subsumed by poetry here
# https://python-poetry.org/docs/pyproject/#include-and-exclude
include = []

[[tool.poetry.source]]
name = "papermill-origami"
url = "https://pypi.org"

[tool.poetry.dependencies]
python = "^3.8"
noteable-origami = "^0.0.5"
papermill = "^2.4.0"
dagstermill = {version = "^1.0.5", optional = true}
cloudpickle = "^2.2.0"

[tool.poetry.dev-dependencies]
flake8-docstrings = "^1.6.0"
pytest-httpx = "^0.21.0"
black = {version = "^22.3.0", allow-prereleases = true}
isort = "^5.10.1"
boto = "^2.49.0"
notebook = "^6.4.11"
pytest = "^7.1.2"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.19.0"
nox = "^2022.1.7"
nox-poetry = "^1.0.0"
pytest-mock = "^3.8.2"
bump2version = "^1.0.1"

[tool.poetry.extras]
dagster = ["dagstermill"]

[tool.poetry.plugins."papermill.engine"]
noteable = "papermill_origami.engine:NoteableEngine"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | src/nbformat

  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
skip-string-normalization = true

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
known_third_party = []

[tool.coverage.run]
branch = false
omit = ["papermill_origami/_version.py", "*/tests/*"]

[tool.coverage.report]
exclude_lines = ["if self.debug:",
                 "pragma: no cover",
                 "raise AssertionError",
                 "raise NotImplementedError",
                 "if __name__ == '__main__':"]
ignore_errors = true
omit = ["papermill_origami/_version.py"]
